tests/requirements.txt
# Pytest for running the tests.
pytest
pytest-django
